+++ tests/acceptance_tests/test_backtest.py
@@ -259,7 +259,7 @@ class TestBacktestAcceptanceTestsGBPUSDBarsInternal:
             trailing_atr_multiple=3.0,
             trailing_offset_type="PRICE",
             trailing_offset=Decimal("0.01"),
-            trigger_type="LAST_TRADE",
+            trigger_type="LAST_PRICE",
         )
         strategy = EMACrossStopEntry(config=config)
         self.engine.add_strategy(strategy)